ENH: Bump sphinx version, add pandoc version, bump RTD python version.
Old sphinx version spec is very old.  Both RTD and CI interpret sphinx>=2.0.0 as
5.0.3, so it seems reasonable that requiring this for developers would only
serve to prevent doc build mistakes.

Also, added pandoc as a dev requirement, because it is a requirement for
building docs.  It has an external binary requirement (pandoc) which can't be
accomodated in a requirements file.  So I added info to the readme.rst

Bump python version to 3.8 in RTD since 3.7 is approaching EOL

FWIW, was confused about pandoc so added an issue to RTFD:
https://github.com/readthedocs/readthedocs.org/issues/10229