CLN: No longer handle unbiasing 1-D self.coef_
It's never hit in our tests, and setting up a test case is pretty contrived,
even more so in the real world.  At the end of the day, the only error it would
cause is in SINDy.equations().  If we want to provide optimizer developers this
safety, I would assume it would be in a utility function (e.g.
`check_optimizer()`), like sklearn's `check_estimator()`.